code refactoring, menu add disabled buttons
