--- src/bundle/Resources/config/routing.yml
@@ -1,5 +1,0 @@
-gie_ez_toolbar_homepage:
-    path:     /toolbar/{locationId}
-    methods: [POST]
-    defaults: { _controller: GieEzToolbarBundle:Toolbar:render }
-
