+++ src/bundle/Resources/config/services.yml
@@ -4,18 +4,6 @@ services:
     autoconfigure: true
     public: false
 
-  Gie\EzToolbarBundle\Controller\ToolbarController:
-    public: true
-    arguments:
-      $languages: '$languages$'
-      $actionDispatcher: "@ezrepoforms.action_dispatcher.content"
-
-  Gie\EzToolbar\Form\Type\ToolbarType:
-    arguments:
-      $contentTypeChoiceLoader: '@EzSystems\EzPlatformAdminUi\Form\Type\ChoiceList\Loader\ContentTypeChoiceLoader'
-    tags:
-      - { name: form.type }
-
   Gie\EzToolbar\Form\Event\UrlRedirectProcessor:
     decorates: EzSystems\RepositoryForms\Form\Processor\SystemUrlRedirectProcessor
     arguments:
